corrigo o merge feito indevidamente

+++ app/Http/Controllers/Api/UserController.php
@@ -2,6 +2,8 @@
 
 namespace App\Http\Controllers\Api;
 
+use App\Http\Requests\UpdateUserRequest;
+use App\Http\Requests\CreateUserRequest;
 use App\Http\Controllers\Controller;
 use App\Models\User;
 use Illuminate\Http\Request;
@@ -15,54 +17,30 @@ class UserController extends Controller
 
     }
 
-    public function create(Request $request)
+    public function create(CreateUserRequest $request)
     {
+        return User::create($request->all());
+    }
 
-        $user = new User();
-
-        $user = $this->setData($user, $request);
-
-        $user->save();
-
+    public function update(UpdateUserRequest $request, User $user)
+    {
+        $user->update($request->all());
         return $user;
     }
 
-    public function update(Request $request)
+    public function delete(Request $request, $id)
     {
-        $user = new User();
+        $user = User::find($id);
 
-        $user = $user->find($request->id);
-
-        $user = $this->setData($user, $request);
-
-        $user->save();
-
-        return $user;
+        if ($user) {
+            $user->delete();
+            return redirect()->route('users.index')->with('success', 'Usuário excluído com sucesso.');
+        } else {
+            return redirect()->route('users.index')->with('error', 'Usuário não encontrado.');
+        }
     }
 
-    public function delete(Request $request)
-    {
-        $user = new User();
 
-        $user = $user->find($request->id);
-
-        $user->delete();
-
-    }
-
-    protected function setData(User $user, Request $request)
-    {
-        $user->nome = $request->nome;
-        $user->email = $request->email;
-        $user->data_nascimento = $request->data_nascimento;
-        $user->cpf = $request->cpf;
-        $user->foto = $request->foto;
-        $user->login = $request->login;
-        $user->senha = $request->senha;
-        $user->perfil = $request->perfil;
-
-        return $user;
-    }
 
 
 }
